requirements.txt
websockets
aioquic
icmplib
tabulate
requests
pysocks
ssl
